修改客户反馈问题

+++ app/src/main/java/com/morrigan/m/main/RankResult.java
@@ -14,5 +14,5 @@ public class RankResult extends HttpResult {
     public int rank;
 
     @Keep
-    public double eValue;
+    public int eValue;
 }
